refactor: components/featuers/ 폴더구조 수정
